% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_pedigree.R
\name{analyse_pedigree}
\alias{analyse_pedigree}
\title{Analyse a pedigree problem}
\usage{
analyse_pedigree(x)
}
\arguments{
\item{x}{Object of class \code{ped}, usually created using the individual
functions \code{\link[=sim_pedigree]{sim_pedigree()}} and \code{\link[=sim_trait]{sim_trait()}}, \code{\link[=sim_trait_xd]{sim_trait_xd()}}, or
\code{\link[=sim_trait_xr]{sim_trait_xr()}}.}
}
\value{
Returns an object of class \code{ped}.

The following list of constellations and characteristics is
monitored and represents the basic framework of the analysis.
\itemize{
\item Affected female with unaffected parents: excludes AD, XD, XR
\item Affected male with unaffected parents: excludes AD, XD
\item Unaffected female with affected parents: excludes AR, XD, XR
\item Unaffected male with affected parents: excludes AR, XR
\item Affected male with unaffected mother: excludes XD
\item Unaffected female with affected father: excludes XD
\item Affected female with unaffected father: excludes XR
\item Unaffected  male with affected mother: excludes XR
\item All daughters, but no sons of affected males are affected as well:
likely XD, unlikely AD
\item The number of independent alleles required differs noticeably for
the two recessive modes of inheritance: likely XR, unlikely AR
}

The linkage of gender and affection status among children of affected
fathers is checked by Fisher's exact test. A significant relationship
indicates that an X-linked dominant inheritance is more likely than an
autosomal dominant inheritance.

Whether the number of independent alleles required for the two recessive
modes of inheritance differs significantly enough in favor of
X-linked recessive inheritance is approximated using a simple
statistical test described by Pocock (2006).
}
\description{
\code{\link[=analyse_pedigree]{analyse_pedigree()}} analyses a simulated pedigree problem to identify the
present mode of inheritance. Pedigrees created using \code{\link[=sim_pedigree_problem]{sim_pedigree_problem()}}
are analysed during simulation. Therefore, calling \code{\link[=analyse_pedigree]{analyse_pedigree()}} by
hand is usually not required.
}
\examples{
set.seed(7)
obj1 <- sim_pedigree()
obj1 <- sim_trait(obj1, inheritance = "AR")
obj1 <- analyse_pedigree(obj1)
obj1$analysis

obj2 <- sim_pedigree_problem(seed = 13)
obj2 <- analyse_pedigree(obj2)
obj2$analysis

}
\references{
Pocock, S. J. (2006). The simplest statistical test: how to check for a
difference between treatments. \emph{BMJ, 332}, 1256-1258.
\url{https://doi.org/10.1136/bmj.332.7552.1256}
}
\seealso{
\code{\link[=sim_trait]{sim_trait()}} and related functions \code{\link[=sim_trait_xd]{sim_trait_xd()}} and
\code{\link[=sim_trait_xr]{sim_trait_xr()}} to simulate the distribution of a trait for a pedigree
simulated using \code{\link[=sim_pedigree]{sim_pedigree()}}.
}
