% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_argument.R
\name{check_argument}
\alias{check_argument}
\title{Check an argument regarding the mode of inheritance}
\usage{
check_argument(x, argument, elaborated_feedback = TRUE, l = c("en", "de"))
}
\arguments{
\item{x}{Object of class \code{ped}, usually created using the
\code{\link[=sim_pedigree_problem]{sim_pedigree_problem()}} function.}

\item{argument}{List containing one claim plus proof. Each claim
needs to consist of one qualifier - confirmed, likely, unlikely,
excluded - plus inheritance in the two-letter code - AD, AR, XD, XR - i.e.
\code{excludedAR}. Possible options for proof are: \code{const}, \code{few_affected},
\code{many_affected}, \code{m_and_f}, \code{mainly_m}, \code{mainly_f}, \code{every_gen},
\code{not_every_gen}, \code{no_male_to_male}, \code{carrier_frequency_low},
\code{carrier_frequency_high}, \code{other_arguments}.
In addition, the list must also include variables father, mother, and child
for specifying a family constellation. If not needed, these
variables can be set to \code{NA}. For consistency reasons, each argument should
also be given an \code{id}, which is used in the app and necessary for
\code{\link[=check_argumentation]{check_argumentation()}} to work. See the examples section to see a complete
argument.}

\item{elaborated_feedback}{If \code{FALSE}, a verbal feedback about the
correctness of the result is generated, but a detailed explanation is
omitted.}

\item{l}{String to specify the feedback's language.
Currently \code{de} for German and \code{en} for English feedback are supported.}
}
\value{
Returns the argument including feedback and a scoring as list.
}
\description{
\code{\link[=check_argument]{check_argument()}} is intended for internal use by the learning app and is
called via \code{\link[=check_argumentation]{check_argumentation()}}. Nevertheless, it can also be used outside
of the app. Supply an argument to retrieve feedback and a scoring.
}
\examples{
obj1 <- sim_pedigree_problem(inheritance = "AD", seed = 1)
argument <- list(id = 1,
                 claim = "excludedAR",
                 proof = "const",
                 father = 12,
                 mother = 13,
                 child = 14)

check_argument(obj1, argument)

}
\seealso{
\code{\link[=check_argumentation]{check_argumentation()}}, which evaluates a full argumentation
consisting of any number of arguments.
}
