% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_argumentation.R
\name{check_argumentation}
\alias{check_argumentation}
\title{Check an entire argumentation regarding the mode of inheritance}
\usage{
check_argumentation(
  x,
  argumentation,
  elaborated_feedback = TRUE,
  l = c("en", "de")
)
}
\arguments{
\item{x}{Object of class \code{ped}, usually created using the
\code{\link[=sim_pedigree_problem]{sim_pedigree_problem()}} function.}

\item{argumentation}{List of individual arguments. See
\code{\link[=check_argument]{check_argument()}} for details.}

\item{elaborated_feedback}{Set to \code{TRUE} to return comprehensive feedback on
the content.}

\item{l}{String to specify the feedback's language.
Currently 'de' for German and 'en' for English feedback are supported.}
}
\value{
Returns a list containing the argumentation including feedback and a
scoring per argument and a table which provides an overview of the
argumentation.
}
\description{
\code{\link[=check_argumentation]{check_argumentation()}} is intended for internal use by the learning app.
Nevertheless, it can also be used outside of the app. Supply an argumentation
for evaluation. Provides feedback and a table with an overall summary of the
argumentation.
}
\examples{
obj1 <- sim_pedigree_problem(inheritance = "AD", seed = 1)
argument1 <- list(id = 1,
                  claim = "excludedAR",
                  proof = "const",
                  father = 12,
                  mother = 13,
                  child = 14)

argument2 <- list(id = 2,
                  claim = "excludedXR",
                  proof = "const",
                  father = 12,
                  mother = 13,
                  child = 14)

argumentation <- list('arg-1' = argument1,
                      'arg-2' = argument2)

check_argumentation(obj1, argumentation)

}
\seealso{
\code{\link[=check_argument]{check_argument()}}, which evaluates an individual argument.
}
