% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedana-package.R
\docType{package}
\name{pedana-package}
\alias{pedana}
\alias{pedana-package}
\title{pedana: A tool for creating pedigree problems and practicing pedigree analysis}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

'pedana' is an R package which at its core provides a learning app for practicing pedigree analysis. In addition, it provides the necessary functions to simulate pedigree problems and to analyze and feedback students' reasoning.
}
\author{
\strong{Maintainer}: Justin Timm \email{justin.timm@uni-due.de} (\href{https://orcid.org/0000-0002-5718-3472}{ORCID})

}
\keyword{internal}
