% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ped.R
\name{plot.ped}
\alias{plot.ped}
\title{Plot simulated pedigrees}
\usage{
\method{plot}{ped}(x, ...)
}
\arguments{
\item{x}{Object of class \code{ped} created by functions \code{\link[=sim_pedigree]{sim_pedigree()}} or
\code{\link[=sim_pedigree_problem]{sim_pedigree_problem()}}}

\item{...}{Other arguments passed to the \code{\link[kinship2:plot.pedigree]{kinship2::plot.pedigree()}} method
from the \code{kinship2} package.}
}
\value{
Returns the invisible list generated by the method plot.pedigree
from the kinship2 package.
}
\description{
Plot method for objects of class \code{ped}. Basically this function is a wrapper
for the \code{\link[kinship2:plot.pedigree]{kinship2::plot.pedigree()}} function from the \code{kinship2} package.
}
\examples{
set.seed(1)
plot(sim_pedigree(generations = 3, tfr = 3))

}
\seealso{
\code{\link[kinship2:plot.pedigree]{kinship2::plot.pedigree()}}
}
