% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pedigree_analysis.R
\name{run_pedigree_analysis}
\alias{run_pedigree_analysis}
\title{Run Shiny application pedana for practicing pedigree analysis}
\usage{
run_pedigree_analysis(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
\code{\link[=run_pedigree_analysis]{run_pedigree_analysis()}} starts the Shiny application \code{pedana} for
practicing pedigree analysis.
}
\examples{
\dontrun{run_pedigree_analysis()}

}
