% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_solution.R
\name{show_solution}
\alias{show_solution}
\title{Show a solution to an existing pedigree problem}
\usage{
show_solution(x, complete = TRUE)
}
\arguments{
\item{x}{Object of class \code{ped}, usually created using the
\code{\link[=sim_pedigree_problem]{sim_pedigree_problem()}}.}

\item{complete}{If \code{TRUE}, show a complete solution.}
}
\value{
Returns a data frame showing a complete or incomplete solution for
an existing pedigree.
}
\description{
\code{\link[=show_solution]{show_solution()}} aggregates the results of the analysis of a pedigree
problem obtained by \code{\link[=analyse_pedigree]{analyse_pedigree()}} and can output a halfway or complete
solution to the given problem.
}
\seealso{
\code{\link[=verbalise_solution]{verbalise_solution()}} to create a text briefly explain a solution
to the problem.
}
