% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pedigree.R
\name{sim_pedigree}
\alias{sim_pedigree}
\title{Simulate a pedigree}
\usage{
sim_pedigree(generations = 4, tfr = 1.5)
}
\arguments{
\item{generations}{Number of \code{generations} to simulate. The minimum number of
\code{generations} is 2. There is no upper limit, but the objects can quickly
become very large for high numbers of \code{generations} and large total fertility
rates (\code{tfr}).}

\item{tfr}{A positive numeric value. The total fertility rate describes how
many children women have on average. Therefore, the \code{tfr} influences the
size of the pedigree.}
}
\value{
Returns an object of class \code{ped}.
}
\description{
\code{\link[=sim_pedigree]{sim_pedigree()}} simulates a pedigree or rather a family structure. A genetic
trait can be simulated afterwards, for example using \code{\link[=sim_trait]{sim_trait()}}. The size
of the pedigree depends largely on the arguments.
}
\examples{
sim_pedigree(generations = 2)
set.seed(1)
plot(sim_pedigree(generations = 3, tfr = 3))

}
\seealso{
\code{\link[=sim_trait]{sim_trait()}} and related functions \code{\link[=sim_trait_xd]{sim_trait_xd()}} and
\code{\link[=sim_trait_xr]{sim_trait_xr()}} to simulate the distribution of a trait for a simulated
pedigree.
}
