% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trait.R
\name{sim_trait}
\alias{sim_trait}
\title{Simulate an inherited trait}
\usage{
sim_trait(
  x,
  inheritance = c("AD", "AR", "XD", "XR"),
  min_affected = 3,
  allele_freq = 0.01
)
}
\arguments{
\item{x}{Object of class \code{ped}, usually created using the
\code{\link[=sim_pedigree]{sim_pedigree()}} function.}

\item{inheritance}{Inheritance of the trait to be simulated.}

\item{min_affected}{Specifies the minimum number of affected individuals to
be simulated.}

\item{allele_freq}{A positive numeric value. With a higher allele frequency,
the chance increases that several independent alleles will be introduced
into the pedigree.}
}
\value{
Returns an object of class \code{ped}.
}
\description{
\code{\link[=sim_trait]{sim_trait()}} simulates an inherited trait for one of the typical modes of
inheritance: autosomal dominant, autosomal recessive, X-linked dominant, or
X-linked recessive. Often the simulation is not successful, especially for
the X-linked modes of inheritance. Use with caution.
}
\examples{
set.seed(4)
obj <- sim_pedigree(generations = 3, tfr = 2)
obj <- sim_trait(obj)
plot(obj)

}
\seealso{
\code{\link[=sim_trait_xd]{sim_trait_xd()}} for X-linked dominant traits and
\code{\link[=sim_trait_xr]{sim_trait_xr()}} for X-linked recessive traits.
}
