% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbalise_solution.R
\name{verbalise_solution}
\alias{verbalise_solution}
\title{Verbalise a solution to an existing pedigree problem}
\usage{
verbalise_solution(x, details = FALSE, l = c("en", "de"))
}
\arguments{
\item{x}{Object of class \code{ped}, usually created using
\code{\link[=sim_pedigree_problem]{sim_pedigree_problem()}}.}

\item{details}{If \code{TRUE}, add evidence to the solution.}

\item{l}{String to specify the feedback's language.
Currently 'de' for German and 'en' for English feedback are supported.}
}
\value{
Returns a string explaining the correct result.
}
\description{
\code{\link[=verbalise_solution]{verbalise_solution()}} creates a text briefly explaining a solution
to the problem in German or English. You can select whether or not to
output evidence to the stated solution.
}
\seealso{
\code{\link[=show_solution]{show_solution()}} aggregates the results of the analysis of a
pedigree problem as data frame.
}
